requirements.txt
flask
numpy
bert-serving-client
termcolor
pyyaml